🔥 Remove and rename SubmissionInfoModal2
