+++ src/routes/moderator/submissions/Components/Indomitable/IndomitableSubmits.svelte
@@ -4,7 +4,7 @@
 	import InfoTooltip from '$lib/Components/InfoTooltip.svelte';
 	import type { DuelRun } from '$lib/types/api/runs/run';
 	import { fetchGetApi } from '$lib/utils/fetch';
-	import SubmissionInfoModal2 from '$lib/Components/SubmissionInfoModal2.svelte';
+	import SubmissionInfoModal2 from '$lib/Components/SubmissionInfoModal.svelte';
 	import type { RunCategories } from '$lib/types/api/categories';
 
 	export let category: RunCategories;
